Moved all the compile time values out of Function
These are all now in the FunctionEntry structure. When the Function is closed the necessary values are copied to the Function. Saves about 100 bytes

+++ src/ExecutionUnit.h
@@ -194,7 +194,7 @@ private:
     Mad<Object> _function;
     Mad<Object> _this;
     uint32_t _localOffset = 0;
-    uint32_t _formalParamCount = 0;
+    uint16_t _formalParamCount = 0;
     uint32_t _actualParamCount = 0;
 
     uint32_t _codeSize;
